+++ src/transformer.js
@@ -2,6 +2,7 @@
  * @typedef { import('./public-types.ts').Coordinates} Coordinates
  * @typedef { ReturnType<typeof parse>[0] } ParseResult;
  */
+import { stripBOM } from "./bom.js";
 import { reverseInnerCoordinates } from "./reverse-inner-coordinates.js";
 import { coordinatesOf } from "./coordinates-of.js";
 import { parse } from "./parse.js";
@@ -53,6 +54,8 @@ export class Transformer {
    * @param {import('./public-types.ts').TransformerOptions} [options]
    */
   constructor(source, options = {}) {
+    source = stripBOM(source);
+
     this.#originalSource = source;
 
     let parsed = parse(source, options);
@@ -405,7 +408,7 @@ export class ParseResultStringUtils {
    * @param {Buffer} buffer
    */
   constructor(buffer) {
-    this.#buffer = buffer;
+    this.#buffer = stripBOM(buffer);
   }
 
   /**
@@ -414,7 +417,7 @@ export class ParseResultStringUtils {
    * @param {ParseResult} parseResult
    */
   contentBefore(parseResult) {
-    return this.#buffer.slice(0, parseResult.range.start).toString();
+    return this.#buffer.subarray(0, parseResult.range.startByte).toString();
   }
 
   /**
@@ -422,7 +425,10 @@ export class ParseResultStringUtils {
    */
   originalContentOf(parseResult) {
     return this.#buffer
-      .slice(parseResult.contentRange.start, parseResult.contentRange.end)
+      .subarray(
+        parseResult.contentRange.startByte,
+        parseResult.contentRange.endByte,
+      )
       .toString();
   }
 
@@ -442,7 +448,10 @@ export class ParseResultStringUtils {
    */
   openingTag(parseResult) {
     let openingTag = this.#buffer
-      .slice(parseResult.startRange.start, parseResult.startRange.end)
+      .subarray(
+        parseResult.startRange.startByte,
+        parseResult.startRange.endByte,
+      )
       .toString();
     return openingTag;
   }
@@ -452,7 +461,7 @@ export class ParseResultStringUtils {
    */
   closingTag(parseResult) {
     let closingTag = this.#buffer
-      .slice(parseResult.endRange.start, parseResult.endRange.end)
+      .subarray(parseResult.endRange.startByte, parseResult.endRange.endByte)
       .toString();
     return closingTag;
   }