+++ tests/transformer.parseResultAt.test.ts
@@ -31,19 +31,19 @@ for (let scenario of coords) {
   it(`matches start ${scenario.start}`, () => {
     const result = t.parseResultAt({ start: scenario.start });
 
-    expect(result?.contentRange?.start).toEqual(scenario.start);
+    expect(result?.contentRange?.startByte).toEqual(scenario.start);
   });
 
   it(`matches end ${scenario.end}`, () => {
     const result = t.parseResultAt({ end: scenario.end });
 
-    expect(result?.contentRange?.end).toEqual(scenario.end);
+    expect(result?.contentRange?.endByte).toEqual(scenario.end);
   });
 
   it(`matches whole thing ${scenario.start}`, () => {
     const result = t.parseResultAt(scenario);
 
-    expect(result?.contentRange?.end).toEqual(scenario.end);
+    expect(result?.contentRange?.endByte).toEqual(scenario.end);
   });
 
   it(`matches line and column ${scenario.line} ${scenario.column}`, () => {
@@ -52,6 +52,6 @@ for (let scenario of coords) {
       column: scenario.column,
     });
 
-    expect(result?.contentRange?.end).toEqual(scenario.end);
+    expect(result?.contentRange?.endByte).toEqual(scenario.end);
   });
 }